change type of distance query with condition kilometers

+++ ncmb-core/src/main/java/com/nifty/cloud/mb/core/NCMBQuery.java
@@ -584,7 +584,7 @@ public class NCMBQuery<T extends NCMBBase> {
      * @param center center location for data searching
      * @param distance search radius distance from center point in kilometers
      */
-    public void whereWithinKilometers(String key, Location center, int distance){
+    public void whereWithinKilometers(String key, Location center, double distance){
         try {
             JSONObject newCondition = new JSONObject();
             if (mWhereConditions.has(key)){
